arreglo de funcionalidades stopwords

+++ src/funciones.php
@@ -43,16 +43,18 @@ function separarPalabras($texto)
     return preg_split('/\s+/u', $texto);
 }
 
-function contarPalabras($palabras)
+function contarPalabras(array $palabras, string $rutaStopwords = __DIR__ . '/stopwords.txt')
 {
-    $stopwords = file('stopwords.txt', FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
-    // Asegura UTF-8 y sin tildes para stopwords
-    foreach ($stopwords as &$stopword) {
-        $stopword = asegurarUTF8($stopword);
-        $stopword = mb_strtolower($stopword, 'UTF-8');
-        $stopword = eliminarTildes($stopword);
+    $stopwords = [];
+    if (file_exists($rutaStopwords)) {
+        $stopwords = file($rutaStopwords, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+        // Normalizar stopwords
+        foreach ($stopwords as &$stopword) {
+            $stopword = asegurarUTF8($stopword);
+            $stopword = mb_strtolower($stopword, 'UTF-8');
+            $stopword = eliminarTildes($stopword);
+        }
     }
-
     $frecuencias = [];
 
     foreach ($palabras as $palabra) {
